+++ src/Core/Language/Utf8/IDocumentCache.cs
@@ -1,0 +1,7 @@
+namespace HotChocolate.Language
+{
+    public interface IDocumentCache
+    {
+        bool TryGet(string key, out DocumentNode document);
+    }
+}
