+++ src/Core/Language/Utf8/IDocumentHashProvider.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace HotChocolate.Language
+{
+    public interface IDocumentHashProvider
+    {
+        string ComputeHash(ReadOnlySpan<byte> document);
+    }
+}
